+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add graph navigator renderer using https://grapheco.github.io/InteractiveGraph/
+
 ### Changed
 - Renderer no longer identified automatically based on json file extension.